requirements.txt
# 核心交易库 (连接 OKX)
ccxt>=4.0.0

# 数据处理与策略计算 (K线与均线计算)
pandas>=2.0.0
numpy>=1.24.0

# 环境变量管理 (读取 .env 文件)
python-dotenv>=1.0.0

# 网络请求 (发送钉钉/TG通知)
requests>=2.31.0

# (可选) 如果你想用更丰富的技术指标库，可以取消下面这行的注释
# pandas-ta

# 可视化与回测
matplotlib>=3.7.0
streamlit>=1.30.0
